# openapi.yaml
openapi: 3.0.1
info:
  description: 'API Todo app'
  version: '1'
  title: 'API Todo app'
servers:
  - url: 'http://localhost:3000'
...
paths:
...
  /todos:
    get:
      tags:
        - 'todos'
      summary: 'Retrieve all todos'
      description: 'Retrieve all todos, only admins can access this ressource'
      operationId: 'getTodos'
      responses:
        '200':
          $ref: '#/components/responses/todosList'
        '401':
          $ref: '#/components/responses/authenticationFailed'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
  
    post:
      tags:
        - 'todos'
      summary: 'Add a new todo'
      description: 'Add a new todo, you need to be authentified'
      operationId: 'addTodo'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addTodo'
      responses:
        '201':
          description: 'Todo added successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/todo'
        '401':
          $ref: '#/components/responses/authenticationFailed'
        '400':
          $ref: '#/components/responses/badRequest'
        '500':
          $ref: '#/components/responses/internalServerError'
  /todos/{todoId}:
    get:
      tags:
        - 'todos'
      summary: 'Retrieve a todo by its id'
      description: 'Retrieve a todo by its id, only admins and owner can access this ressource'
      operationId: 'getTodoById'
      parameters:
        - name: 'todoId'
          in: 'path'
          required: true
          description: 'ID of the todo'
          schema:
            type: 'number'
      responses:
        '200':
          description: 'Todo retrieved successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/todo'
        '401':
          $ref: '#/components/responses/authenticationFailed'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'

    patch:
      tags:
        - 'todos'
      summary: 'Update a todo by ID'
      description: 'Update a todo by ID, only admins and owner can modify this ressource.'
      operationId: 'updateTodoById'
      parameters:
        - name: 'todoId'
          in: 'path'
          required: true
          description: 'ID of the todo to update'
          schema:
            type: 'number'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateTodo'
      responses:
        '200':
          description: 'Todo updated successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/todo'
        '401':
          $ref: '#/components/responses/authenticationFailed'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'

    delete:
      tags:
        - 'todos'
      summary: 'Delete a todo by ID'
      description: 'Delete a todo by ID, only admins and owner can delete this ressource'
      operationId: 'deleteTodoById'
      parameters:
        - name: 'todoId'
          in: 'path'
          required: true
          description: 'ID of the todo to delete'
          schema:
            type: 'number'
      responses:
        '200':
          description: 'Todo deleted successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/todo'
        '401':
          $ref: '#/components/responses/authenticationFailed'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /users/{userId}/todos:
    get:
      tags:
        - 'todos'
      summary: 'Retrieve all todos by user ID'
      description: 'Retrieve all todos of an user by its ID, only admins and corresponding user can access this ressource'
      operationId: 'getTodosByUserId'
      parameters:
        - name: 'userId'
          in: 'path'
          required: true
          description: 'ID of the user'
          schema:
            type: 'number'
      responses:
        '200':
          $ref: '#/components/responses/todosList'
        '401':
          $ref: '#/components/responses/authenticationFailed'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
components:
  schemas:
...
    todo:
      type: 'object'
      properties:
        id:
          type: 'number'
          example: 1
        title:
          type: 'string'
          example: 'Do something'
        ownerId:
          type: 'number'
          example: 1
        done:
          type: 'boolean'
          example: false
      required:
        - id
        - title
        - ownerId
        - done
    addTodo:
      type: 'object'
      properties:
        title:
          type: 'string'
          example: 'Do something'
      required:
        - title
    updateTodo:
      type: 'object'
      properties:
        title:
          type: 'string'
          example: 'Do something else'
        done:
          type: 'boolean'
          example: true
  responses:
...
    notFound:
      description: 'Not Found'
      content:
        application/json:
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
                example: 'Ressource not found'
    badRequest:
      description: 'Bad Request'
      content:
        application/json:
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
                example: 'The reason why request failed'
    internalServerError:
      description: 'Internal Server Error'
      content:
        application/json:
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
                example: 'Internal server error, please try again later'
    authenticationFailed:
      description: 'Authentication failed'
      content:
        application/json:
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
                example: 'Authentication failed'
    forbidden:
      description: 'Access forbidden'
      content:
        application/json:
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
                example: 'Access forbidden'
    todosList:
      description: 'An array of all todos'
      content:
        application/json:
          schema:
            type: 'array'
            items:
              $ref: '#/components/schemas/todo'